1 Compartment Model Simulated Data from ACOP 2016

This is a simulated dataset from the ACOP 2016 poster. All Datasets were simulated with the following methods.

Usage

Bolus_1CPT

Format

A data frame with 7,920 rows and 14 columns

ID

Simulated Subject ID

TIME

Simulated Time

DV

Simulated Dependent Variable

LNDV

Simulated log(Dependent Variable)

MDV

Missing DV data item

AMT

Dosing AMT

EVID

NONMEM Event ID

DOSE

Dose

V

Individual Simulated Volume

CL

Individual Clearance

SS

Steady State

II

Interdose Interval

SD

Single Dose Flag

CMT

Compartment

Source

Schoemaker R, Xiong Y, Wilkins J, Laveille C, Wang W. nlmixr2: an open-source package for pharmacometric modelling in R. ACOP 2016

Details

Richly sampled profiles were simulated for 4 different dose levels (10, 30, 60 and 120 mg) of 30 subjects each as single dose (over 72h), multiple dose (4 daily doses), single and multiple dose combined, and steady state dosing, for a range of test models: 1- and 2-compartment disposition, with and without 1st order absorption, with either linear or Michaelis-Menten (MM) clearance(MM without steady state dosing). This provided a total of 42 test cases. All inter-individual variabilities (IIVs) were set at 30%, residual error at 20% and overlapping PK parameters were the same for all models. A similar set of models was previously used to compare NONMEM and Monolix4. Estimates of population parameters, standard errors for fixed-effect parameters, and run times were compared both for closed-form solutions and using ODEs. Additionally, a sparse data estimation situation was investigated where 500 datasets of 600 subjects each (150 per dose) were generated consisting of 4 random time point samples in 24 hours per subject, using a first-order absorption, 1-compartment disposition, linear elimination model.

See also