One compartment test dataset showing NONMEM 7.4.3 output

This is a example dataset originally created to show how similar mrgsolve and NONMEM were (See ).

Usage

nmtest

Format

A data frame with 7,157 rows and 15 columns

id

NONMEM id

time

NONMEM time

cp

NONMEM cp output from 7.4.3

cmt

cmt specification 1=depot, 2=central

amt

Nonmem dose

evid

NONMEM Event ID

ii

Interdose Interval

ss

Steady state flag

addl

Individual Clearance

rate

Rate of the infusion

lagt

Lag time

bioav

Bioavailability

rat2

Modeled rate when mode == 1

dur2

Duration when mode == 2

mode

Mode = 0 is no modification, modeled rate when mode=1 and modeled duration when mode=2

Details

The original dataset was created by Kyle Baron and is composed of id<100 the id>100 are modifications by Matthew Fidler to benchmark steady state infusions with lag times and other uncommon features.

Note that rxode2/nlmixr2 will not always match these behaviors by default, we choose behaviors that we believe make sense. There are options to make rxode2/nlmixr2 behave more like NONMEM. However behaviors we believe are wrong we do not support.

See also

Author

Kyle Baron & Matthew Fidler